extract commands into testable functions
